requirements.txt
adium-theme-ubuntu==0.3.4
alabaster==0.7.7
astroid==1.4.4
Babel==1.3
backports-abc==0.5
backports.functools-lru-cache==1.4
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.4.1
bleach==2.0.0
certifi==2017.7.27.1
chardet==2.3.0
configparser==3.5.0
cvxopt==1.1.4
cycler==0.9.0
decorator==4.0.6
docutils==0.12
entrypoints==0.2.3
enum34==1.1.6
fandango==13.2.5
fisx==1.1.4
functools32==3.2.3.post2
future==0.15.2
guidata==1.7.4
guiqwt==3.0.2
h5py==2.6.0
html5lib==0.999
idna==2.6
imagesize==0.7.1
ipykernel==4.6.1
ipython==2.4.1
ipython-genutils==0.2.0
isort==4.2.15
itango==0.1.6
jdcal==1.0
jedi==0.9.0
Jinja2==2.8
joblib==0.9.4
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-core==4.3.0
lazy-object-proxy==1.2.1
logilab-common==1.1.0
lxml==3.5.0
MarkupSafe==0.23
matplotlib==1.5.1
mccabe==0.6.1
MDP==3.5
mercurial==3.7.3
mistune==0.7.4
MySQL-python==1.2.5
mysqlclient==1.3.7
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
numexpr==2.4.3
numpy==1.13.3
numpydoc==0.7.0
openpyxl==2.3.0
pandas==0.17.1
pandocfilters==1.4.2
panic==6.4.0
pathlib2==2.3.0
patsy==0.4.1
pexpect==4.0.1
pickleshare==0.7.4
Pillow==3.1.2
ply==3.10
pp==1.6.4
prompt-toolkit==1.0.15
psutil==3.4.2
ptyprocess==0.5
pulsectl==23.5.2
py==1.4.31
pycodestyle==2.3.1
pydicom==0.9.9
pyflakes==1.1.0
Pygments==2.1
pygobject==3.20.0
pylint==1.5.2
PyMca==4.7.4
PyMca5==5.2.1
PyOpenGL==3.0.2
pyparsing==2.0.3
Pyste==0.9.10
pytango==9.2.1
PyTangoArchiving==7.5.0
pytest==2.8.7
python-dateutil==2.4.2
PythonQwt==0.5.5
pytz==2014.10
pyzmq==15.2.0
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
roman==2.0.0
rope==0.10.2
sardana==2.3.2
scandir==1.6
scikit-learn==0.17
scipy==0.17.0
simplegeneric==0.8.1
simplejson==3.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.3.6
sphinx-rtd-theme==0.1.9
sphinxcontrib-websupport==1.0.1
spyder==2.3.8
statsmodels==0.6.1
tables==3.2.2
taurus==4.1.1
testpath==0.3.1
tgconf-SardanaGUI==0.1.0
tornado==4.5.2
traitlets==4.3.2
typing==3.6.2
unity-lens-photos==1.0
urllib3==1.22
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5.1
wrapt==1.8.0
xlrd==0.9.4
xlwt==0.7.5
